Fiji implementation

+++ app/src/main/java/com/alphawallet/app/service/AlphaWalletService.java
@@ -187,7 +187,7 @@ public class AlphaWalletService
         {
             args.put("r", Numeric.toHexString(sigData.getR()));
             args.put("s", Numeric.toHexString(sigData.getS()));
-            args.put("v", Integer.toHexString(sigData.getV()));
+            args.put("v", Numeric.toHexString(sigData.getV()));
         }
     }
